[probes] added extensive testing for core ops in both rust and python

+++ arrgen/src/lib.rs
@@ -4,7 +4,7 @@ use numpy::{IxDyn, PyArrayDyn, ToPyArray};
 use pyo3::prelude::*;
 use rand::{distributions::Uniform, rngs::StdRng, SeedableRng};
 
-pub fn uniform_array(seed: u64, shape: Vec<usize>, min: f64, max: f64) -> ArrayD<f64> {
+pub fn uniform_array(seed: u64, shape: Vec<usize>, min: f32, max: f32) -> ArrayD<f32> {
     let mut rng = StdRng::seed_from_u64(seed);
     let dist = Uniform::new(min, max);
     let array_shape = IxDyn(&shape);
@@ -12,7 +12,7 @@ pub fn uniform_array(seed: u64, shape: Vec<usize>, min: f64, max: f64) -> ArrayD
     array
 }
 
-pub fn normal_array(seed: u64, shape: Vec<usize>, mean: f64, std_dev: f64) -> ArrayD<f64> {
+pub fn normal_array(seed: u64, shape: Vec<usize>, mean: f32, std_dev: f32) -> ArrayD<f32> {
     let mut rng = StdRng::seed_from_u64(seed);
     let dist = Normal::new(mean, std_dev).unwrap();
     let array_shape = IxDyn(&shape);
@@ -20,7 +20,7 @@ pub fn normal_array(seed: u64, shape: Vec<usize>, mean: f64, std_dev: f64) -> Ar
     array
 }
 
-pub fn constant_array(shape: Vec<usize>, value: f64) -> ArrayD<f64> {
+pub fn constant_array(shape: Vec<usize>, value: f32) -> ArrayD<f32> {
     let array_shape = IxDyn(&shape);
     ArrayD::from_elem(array_shape, value)
 }
@@ -31,9 +31,9 @@ fn py_uniform_array(
     py: Python<'_>,
     seed: u64,
     shape: Vec<usize>,
-    min: f64,
-    max: f64,
-) -> PyResult<Bound<'_, PyArrayDyn<f64>>> {
+    min: f32,
+    max: f32,
+) -> PyResult<Bound<'_, PyArrayDyn<f32>>> {
     let array = uniform_array(seed, shape, min, max);
     Ok(array.to_pyarray(py))
 }
@@ -44,9 +44,9 @@ fn py_normal_array(
     py: Python<'_>,
     seed: u64,
     shape: Vec<usize>,
-    mean: f64,
-    std_dev: f64,
-) -> PyResult<Bound<'_, PyArrayDyn<f64>>> {
+    mean: f32,
+    std_dev: f32,
+) -> PyResult<Bound<'_, PyArrayDyn<f32>>> {
     let array = normal_array(seed, shape, mean, std_dev);
     Ok(array.to_pyarray(py))
 }
@@ -56,8 +56,8 @@ fn py_normal_array(
 fn py_constant_array(
     py: Python<'_>,
     shape: Vec<usize>,
-    value: f64,
-) -> PyResult<Bound<'_, PyArrayDyn<f64>>> {
+    value: f32,
+) -> PyResult<Bound<'_, PyArrayDyn<f32>>> {
     let array = constant_array(shape, value);
     Ok(array.to_pyarray(py))
 }
